requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
importlib-metadata==3.3.0
iniconfig==1.1.1
more-itertools==8.6.0
MouseInfo==0.1.3
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pytest==4.6.10
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
PyTweening==1.0.3
pywin32==300
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
